updated unhvd to send aux frame 1 PCM data to AAudio as int16

+++ unhvd-native-android/aaos.h
@@ -14,7 +14,7 @@ struct aaos* aaos_init();
 void aaos_close(struct aaos* a);
 
 // push the samples directly to the buffer
-int32_t aaos_write(struct aaos* a, const float* buffer, const int32_t buflen);
+int32_t aaos_write(struct aaos* a, const int16_t* buffer, const int32_t buflen);
 
 #ifdef __cplusplus
 }
